Last Commit

+++ MCP_Domain/Models/Scooter.cs
@@ -8,7 +8,7 @@ namespace MCP_Domain.Models
     {
         public int Id { get; set; }
         public string Brand { get; set; } = string.Empty;
-        public float Battery_Capacity { get; set; }
+        public double Battery_Capacity { get; set; }
         public Status? Status { get; set; }
         public IEnumerable<Trip> Trips { get; set; }
         public Scooter (string brand)
